// admin/src/api/generated/schemas/adminAccountingLedgerPostingItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminAccountingLedgerPostingItem = {
  seq?: number;
  asset_code?: string;
  scale?: number;
  bucket?: string;
  account_id?: string;
  owner_type?: string;
  owner_id?: string;
  account_type_code?: string;
  account_type_name?: string;
  account_type_description?: string;
  account_normal_side?: string;
  chain_scope?: string;
  debit_raw?: string;
  credit_raw?: string;
  debit_decimal?: string;
  credit_decimal?: string;
};