// admin/src/api/generated/schemas/adminAccountingLedgerTxItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminAccountingLedgerTxItem = {
  tx_id?: string;
  created_at?: string;
  op_type?: string;
  biz_ref?: string;
  idempotency_key?: string;
  request_hash?: string;
};